refactor: refactor fetch functions
